+++ app/models/admin/resource.rb
@@ -2,13 +2,21 @@
 
 # Defines resources for MyApnea website.
 class Admin::Resource < ApplicationRecord
+  # Concerns
+  include Deletable
+  include Sluggable
+
   # Validations
   validates :name, presence: true
   validates :position, presence: true
 
+
   # Callback
   mount_uploader :photo, PhotoUploader
 
-  # Concerns
-  include Deletable
+  # Methods
+  def destroy
+    update slug: nil
+    super
+  end
 end