+++ db/migrate/20220417192557_add_slug_to_admin_resources.rb
@@ -1,0 +1,6 @@
+class AddSlugToAdminResources < ActiveRecord::Migration[6.1]
+  def change
+    add_column :admin_resources, :slug, :string
+    add_index :admin_resources, :slug, unique: true
+  end
+end